requirements.txt
streamlit>=1.32.0
streamlit-extras>=0.4.0
google-genai>=0.2.0
groq>=0.11.0
pdfplumber>=0.10.0
plotly>=5.18.0
python-dotenv>=1.0.0
reportlab>=4.0.0
Pillow>=10.3.0
wordcloud>=1.9.0
textstat>=0.7.0
pandas>=2.0.0
openpyxl>=3.1.0
